[refactor] changepost폴더 하위 페이지들 이름 변경
- 기존 동사형에서 명사형으로 변경
